fix caller dtype mismatch

+++ examples/aot/fast_inverse/caller.cpp
@@ -18,8 +18,8 @@ extern "C" void call_kernel(
 {
     KERNEL_FN<<<blockDim, nullptr, stream>>>(
         reinterpret_cast<float *>(tensor_out),
-        reinterpret_cast<float *>(tensor_in),
-        reinterpret_cast<float *>(identity_in),
+        reinterpret_cast<half *>(tensor_in),
+        reinterpret_cast<half *>(identity_in),
         static_cast<int32_t>(matrix_size),
         static_cast<int32_t>(max_block_size));
 }
